% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioImport.R
\name{bioImport}
\alias{bioImport}
\title{bioImport}
\usage{
bioImport(x, group = "mi", dfref = NULL)
}
\arguments{
\item{x}{user \code{data.frame}.}

\item{group}{\code{mi} for macroinvertebrates and \code{mf} for macrophytes.}

\item{dfref}{custom reference database.}
}
\description{
bioImport
}
